+++ src/context/build_router.rs
@@ -7,7 +7,6 @@ use tower_http::timeout::TimeoutLayer;
 /// A builder to construct `axum::Route` of CometD server.
 #[derive(Debug)]
 pub struct RouterBuilder {
-    base_path: &'static str,
     subscribe_base_path: &'static str,
     handshake_base_path: &'static str,
     connect_base_path: &'static str,
@@ -18,11 +17,10 @@ impl Default for RouterBuilder {
     #[inline(always)]
     fn default() -> Self {
         Self {
-            base_path: "/",
             subscribe_base_path: "/",
-            handshake_base_path: "/",
-            connect_base_path: "/",
-            disconnect_base_path: "/",
+            handshake_base_path: "",
+            connect_base_path: "",
+            disconnect_base_path: "",
         }
     }
 }
@@ -49,60 +47,30 @@ impl RouterBuilder {
         Msg: Debug + Clone + Send + Serialize + 'static,
     {
         let Self {
-            base_path,
             subscribe_base_path,
             handshake_base_path,
             connect_base_path,
             disconnect_base_path,
         } = self;
 
-        let subscribe_route = Router::new().route(subscribe_base_path, post(subscribe::<Msg>));
-        let handshake_route = Router::new().route(
-            &format!("{handshake_base_path}/handshake"),
-            post(handshake::<Msg>),
-        );
-        let connect_route = Router::new().route(
-            &format!("{connect_base_path}/connect"),
-            post(connect::<Msg>),
-        );
-        let disconnect_route = Router::new().route(
-            &format!("{disconnect_base_path}/disconnect"),
-            post(disconnect::<Msg>),
-        );
-
         Router::new()
-            .nest(
-                base_path,
-                Router::new()
-                    .merge(subscribe_route)
-                    .merge(handshake_route)
-                    .merge(connect_route)
-                    .merge(disconnect_route)
-                    .with_state(context.clone()),
+            .route(subscribe_base_path, post(subscribe::<Msg>))
+            .route(
+                &format!("{handshake_base_path}/handshake"),
+                post(handshake::<Msg>),
             )
+            .route(
+                &format!("{connect_base_path}/connect"),
+                post(connect::<Msg>),
+            )
+            .route(
+                &format!("{disconnect_base_path}/disconnect"),
+                post(disconnect::<Msg>),
+            )
+            .with_state(context.clone())
             .layer(TimeoutLayer::new(Duration::from_millis(
                 context.consts.timeout_ms,
             )))
-    }
-
-    /// Set root base-path for routers.
-    ///
-    /// # Example
-    /// ```rust
-    /// use axum_cometd::RouterBuilder;
-    ///
-    /// # let context = axum_cometd::LongPoolingServiceContextBuilder::new().build::<()>();
-    /// let app = RouterBuilder::new()
-    ///     // Ex: `/handshake` -> `/foo/handshake`
-    ///     .base_path("/foo/")
-    ///     .build(&context);
-    /// ```
-    #[inline(always)]
-    pub fn base_path(self, path: &'static str) -> Self {
-        Self {
-            base_path: path,
-            ..self
-        }
     }
 
     /// Set subscribe base-path for routers.
@@ -113,9 +81,8 @@ impl RouterBuilder {
     ///
     /// # let context = axum_cometd::LongPoolingServiceContextBuilder::new().build::<()>();
     /// let app = RouterBuilder::new()
-    ///     .base_path("/foo/")
-    ///     // Ex: `/foo` -> `/foo/bar`
-    ///     .subscribe_base_path("/bar/")
+    ///     // Ex: `/` -> `/bar`
+    ///     .subscribe_base_path("/bar")
     ///     .build(&context);
     /// ```
     #[inline(always)]
@@ -134,9 +101,8 @@ impl RouterBuilder {
     ///
     /// # let context = axum_cometd::LongPoolingServiceContextBuilder::new().build::<()>();
     /// let app = RouterBuilder::new()
-    ///     .base_path("/foo/")
-    ///     // Ex: `/foo/handshake` -> `/foo/bar/handshake`
-    ///     .handshake_base_path("/bar/")
+    ///     // Ex: `/handshake` -> `/bar/handshake`
+    ///     .handshake_base_path("/bar")
     ///     .build(&context);
     /// ```
     #[inline(always)]
@@ -155,9 +121,8 @@ impl RouterBuilder {
     ///
     /// # let context = axum_cometd::LongPoolingServiceContextBuilder::new().build::<()>();
     /// let app = RouterBuilder::new()
-    ///     .base_path("/foo/")
-    ///     // Ex: `/foo/connect` -> `/foo/bar/connect`
-    ///     .connect_base_path("/bar/")
+    ///     // Ex: `/connect` -> `/bar/connect`
+    ///     .connect_base_path("/bar")
     ///     .build(&context);
     /// ```
     #[inline(always)]
@@ -176,9 +141,8 @@ impl RouterBuilder {
     ///
     /// # let context = axum_cometd::LongPoolingServiceContextBuilder::new().build::<()>();
     /// let app = RouterBuilder::new()
-    ///     .base_path("/foo/")
-    ///     // Ex: `/foo/disconnect` -> `/foo/bar/disconnect`
-    ///     .disconnect_base_path("/bar/")
+    ///     // Ex: `/disconnect` -> `/bar/disconnect`
+    ///     .disconnect_base_path("/bar")
     ///     .build(&context);
     /// ```
     #[inline(always)]